requirements.txt
fastapi
uvicorn
google-adk
reportlab
python-dotenv 